+++ packages/extension-base/src/services/swap-service/handler/chainflip-handler.ts
@@ -2,13 +2,14 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import { SwapSDK } from '@chainflip/sdk/swap';
+import { COMMON_ASSETS } from '@subwallet/chain-list';
 import { _ChainAsset } from '@subwallet/chain-list/types';
 import { SwapError } from '@subwallet/extension-base/background/errors/SwapError';
 import { TransactionError } from '@subwallet/extension-base/background/errors/TransactionError';
 import { BasicTxErrorType, ChainType, ExtrinsicType } from '@subwallet/extension-base/background/KoniTypes';
 import { createTransferExtrinsic } from '@subwallet/extension-base/koni/api/dotsama/transfer';
 import { getERC20TransactionObject, getEVMTransactionObject } from '@subwallet/extension-base/koni/api/tokens/evm/transfer';
-import { _getAssetDecimals, _getContractAddressOfToken, _getTokenMinAmount, _isNativeToken, _isSubstrateChain } from '@subwallet/extension-base/services/chain-service/utils';
+import { _getAssetDecimals, _getChainNativeTokenSlug, _getContractAddressOfToken, _getTokenMinAmount, _isNativeToken, _isSubstrateChain } from '@subwallet/extension-base/services/chain-service/utils';
 import { SwapBaseHandler, SwapBaseHandlerInitParams, SwapBaseInterface } from '@subwallet/extension-base/services/swap-service/handler/base-handler';
 import { calculateSwapRate, CHAIN_FLIP_SUPPORTED_MAINNET_ASSET_MAPPING, CHAIN_FLIP_SUPPORTED_MAINNET_MAPPING, CHAIN_FLIP_SUPPORTED_TESTNET_ASSET_MAPPING, CHAIN_FLIP_SUPPORTED_TESTNET_MAPPING, DEFAULT_SWAP_FIRST_STEP, getSwapEarlyValidationError, MOCK_SWAP_FEE, SWAP_QUOTE_TIMEOUT_MAP } from '@subwallet/extension-base/services/swap-service/utils';
 import { TransactionData, YieldStepType } from '@subwallet/extension-base/types';
@@ -17,7 +18,6 @@ import { AxiosError } from 'axios';
 import BigNumber from 'bignumber.js';
 
 import { SubmittableExtrinsic } from '@polkadot/api/types';
-import {COMMON_ASSETS} from "@subwallet/chain-list";
 
 enum ChainflipFeeType {
   INGRESS = 'INGRESS',
@@ -141,16 +141,8 @@ export class ChainflipSwapHandler implements SwapBaseInterface {
       if (srcAssetData.maximumSwapAmount) {
         const bnMaxProtocolSwap = new BigNumber(srcAssetData.maximumSwapAmount);
 
-        console.log('Max swap: ', bnMaxProtocolSwap.shiftedBy(-_getAssetDecimals(fromAsset)).toString());
-
         bnSwapMaxAllowance = BigNumber.min(bnMaxProtocolSwap, bnMaxBalanceSwap);
-      } else {
-        console.log('Max swap: null');
       }
-
-      console.log('Max balance swap: ', bnSwapMaxAllowance.shiftedBy(-_getAssetDecimals(fromAsset)).toString());
-      console.log('Min swap: ', bnMinSwap.shiftedBy(-_getAssetDecimals(fromAsset)).toString());
-      console.log('Max swap allow: ', bnSwapMaxAllowance.shiftedBy(-_getAssetDecimals(fromAsset)).toString());
 
       if (bnMinSwap.gte(bnSwapMaxAllowance)) {
         return {
@@ -230,7 +222,7 @@ export class ChainflipSwapHandler implements SwapBaseInterface {
   }
 
   private parseSwapPath (fromAsset: _ChainAsset, toAsset: _ChainAsset) {
-    if (toAsset.slug !== this.intermediaryAssetSlug) { // Chainflip always use USDC as intermediary
+    if (toAsset.slug !== this.intermediaryAssetSlug && fromAsset.slug !== this.intermediaryAssetSlug) { // Chainflip always use USDC as intermediary
       return [fromAsset.slug, this.intermediaryAssetSlug, toAsset.slug]; // todo: generalize this
     }
 
@@ -241,13 +233,14 @@ export class ChainflipSwapHandler implements SwapBaseInterface {
     const fromAsset = this.chainService.getAssetBySlug(request.pair.from);
     const toAsset = this.chainService.getAssetBySlug(request.pair.to);
 
+    const fromChain = this.chainService.getChainInfoByKey(fromAsset.originChain);
+    const fromChainNativeTokenSlug = _getChainNativeTokenSlug(fromChain);
+
     if (!fromAsset || !toAsset) {
       return new SwapError(SwapErrorType.UNKNOWN);
     }
 
     const earlyValidation = await this.validateSwapRequest(request);
-
-    console.log('Error: ', earlyValidation.error);
 
     const metadata = earlyValidation.metadata as ChainflipPreValidationMetadata;
 
@@ -269,8 +262,6 @@ export class ChainflipSwapHandler implements SwapBaseInterface {
         destAsset: toAssetId,
         amount: request.fromAmount
       });
-
-      console.log('quoteResponse', quoteResponse);
 
       const feeComponent: SwapFeeComponent[] = [];
 
@@ -306,6 +297,8 @@ export class ChainflipSwapHandler implements SwapBaseInterface {
         }
       });
 
+      const defaultFeeToken = _isNativeToken(fromAsset) ? fromAsset.slug : fromChainNativeTokenSlug;
+
       return {
         pair: request.pair,
         fromAmount: request.fromAmount,
@@ -319,8 +312,8 @@ export class ChainflipSwapHandler implements SwapBaseInterface {
         isLowLiquidity: quoteResponse.quote.lowLiquidityWarning,
         feeInfo: {
           feeComponent: feeComponent,
-          defaultFeeToken: fromAsset.slug,
-          feeOptions: [fromAsset.slug]
+          defaultFeeToken,
+          feeOptions: [defaultFeeToken]
         },
         route: {
           path: this.parseSwapPath(fromAsset, toAsset)
@@ -330,7 +323,11 @@ export class ChainflipSwapHandler implements SwapBaseInterface {
       const error = e as AxiosError;
       const errorObj = error?.response?.data as Record<string, string>;
 
-      if (errorObj.error.includes(CHAINFLIP_QUOTE_ERROR.InsufficientLiquidity)) {
+      if (errorObj && errorObj.error && errorObj.error.includes(CHAINFLIP_QUOTE_ERROR.InsufficientLiquidity)) { // todo: Chainflip will improve this
+        return new SwapError(SwapErrorType.NOT_ENOUGH_LIQUIDITY);
+      }
+
+      if (errorObj && errorObj.message && errorObj.message.includes(CHAINFLIP_QUOTE_ERROR.InsufficientLiquidity)) {
         return new SwapError(SwapErrorType.NOT_ENOUGH_LIQUIDITY);
       }
 
@@ -398,8 +395,6 @@ export class ChainflipSwapHandler implements SwapBaseInterface {
       destAddress: receiver,
       amount: quote.fromAmount
     });
-
-    console.log('depositAddressResponse', depositAddressResponse);
 
     const txData: ChainflipTxData = {
       address,