+++ packages/extension-base/src/services/transaction-service/utils.ts
@@ -31,6 +31,10 @@ function getBlockExplorerAccountRoute (explorerLink: string) {
     return 'account';
   }
 
+  if (explorerLink.includes('explorer.polimec.org')) {
+    return 'account';
+  }
+
   return 'address';
 }