+++ packages/extension-web-ui/src/Popup/Transaction/variants/Earn.tsx
@@ -3,32 +3,35 @@
 
 import { _ChainAsset } from '@subwallet/chain-list/types';
 import { ExtrinsicType, NotificationType } from '@subwallet/extension-base/background/KoniTypes';
+import { getValidatorLabel } from '@subwallet/extension-base/koni/api/staking/bonding/utils';
 import { _getAssetDecimals, _getAssetSymbol, _getSubstrateGenesisHash, _isChainEvmCompatible } from '@subwallet/extension-base/services/chain-service/utils';
+import { _STAKING_CHAIN_GROUP } from '@subwallet/extension-base/services/earning-service/constants';
 import { isLendingPool, isLiquidPool } from '@subwallet/extension-base/services/earning-service/utils';
 import { SWTransactionResponse } from '@subwallet/extension-base/services/transaction-service/types';
-import { NominationPoolInfo, OptimalYieldPath, OptimalYieldPathParams, SubmitJoinNativeStaking, SubmitJoinNominationPool, SubmitYieldJoinData, ValidatorInfo, YieldPoolInfo, YieldPoolType, YieldStepType } from '@subwallet/extension-base/types';
+import { EarningStatus, NominationPoolInfo, OptimalYieldPath, OptimalYieldPathParams, SubmitJoinNativeStaking, SubmitJoinNominationPool, SubmitYieldJoinData, ValidatorInfo, YieldPoolInfo, YieldPoolType, YieldStepType } from '@subwallet/extension-base/types';
 import { addLazy } from '@subwallet/extension-base/utils';
 import { AccountSelector, AlertBox, AmountInput, EarningPoolSelector, EarningValidatorSelector, HiddenInput, InfoIcon, LoadingScreen, MetaInfo } from '@subwallet/extension-web-ui/components';
 import { EarningProcessItem } from '@subwallet/extension-web-ui/components/Earning';
 import { getInputValuesFromString } from '@subwallet/extension-web-ui/components/Field/AmountInput';
 import { EarningInstructionModal } from '@subwallet/extension-web-ui/components/Modal/Earning';
-import { BN_ZERO, EARNING_INSTRUCTION_MODAL, STAKE_ALERT_DATA } from '@subwallet/extension-web-ui/constants';
+import { BN_ZERO, EARNING_INSTRUCTION_MODAL, EVM_ACCOUNT_TYPE, STAKE_ALERT_DATA, SUBSTRATE_ACCOUNT_TYPE } from '@subwallet/extension-web-ui/constants';
 import { ScreenContext } from '@subwallet/extension-web-ui/contexts/ScreenContext';
 import { WebUIContext } from '@subwallet/extension-web-ui/contexts/WebUIContext';
-import { useChainConnection, useFetchChainState, useGetBalance, useGetNativeTokenSlug, useInitValidateTransaction, usePreCheckAction, useRestoreTransaction, useSelector, useTransactionContext, useWatchTransaction, useYieldPositionDetail } from '@subwallet/extension-web-ui/hooks';
+import { useChainConnection, useFetchChainState, useGetBalance, useGetNativeTokenSlug, useInitValidateTransaction, usePreCheckAction, useRestoreTransaction, useSelector, useSetSelectedAccountTypes, useTransactionContext, useWatchTransaction, useYieldPositionDetail } from '@subwallet/extension-web-ui/hooks';
 import { insufficientMessages } from '@subwallet/extension-web-ui/hooks/transaction/useHandleSubmitTransaction';
 import { fetchPoolTarget, getOptimalYieldPath, submitJoinYieldPool, validateYieldProcess } from '@subwallet/extension-web-ui/messaging';
 import { unlockDotCheckCanMint } from '@subwallet/extension-web-ui/messaging/campaigns';
 import { DEFAULT_YIELD_PROCESS, EarningActionType, earningReducer } from '@subwallet/extension-web-ui/reducer';
 import { store } from '@subwallet/extension-web-ui/stores';
 import { EarnParams, FormCallbacks, FormFieldData, Theme, ThemeProps } from '@subwallet/extension-web-ui/types';
-import { convertFieldToObject, isAccountAll, parseNominations, reformatAddress, simpleCheckForm } from '@subwallet/extension-web-ui/utils';
+import { convertFieldToObject, getValidatorKey, isAccountAll, parseNominations, reformatAddress, simpleCheckForm } from '@subwallet/extension-web-ui/utils';
 import { ActivityIndicator, Button, ButtonProps, Form, Icon, ModalContext, Number, Typography } from '@subwallet/react-ui';
 import BigN from 'bignumber.js';
 import CN from 'classnames';
-import { CheckCircle, PlusCircle } from 'phosphor-react';
+import { CheckCircle, PlusCircle, XCircle } from 'phosphor-react';
 import React, { useCallback, useContext, useEffect, useMemo, useReducer, useRef, useState } from 'react';
 import { useTranslation } from 'react-i18next';
+import { useNavigate } from 'react-router-dom';
 import { Divider } from 'semantic-ui-react';
 import styled, { useTheme } from 'styled-components';
 
@@ -39,6 +42,7 @@ import { getJoinYieldParams } from '../helper';
 import { EarnOutlet, FreeBalance, FreeBalanceToEarn, TransactionContent, TransactionFooter } from '../parts';
 
 type Props = ThemeProps;
+type ComponentProps = { className?: string; }
 
 const hideFields: Array<keyof EarnParams> = ['slug', 'chain', 'asset'];
 const validateFields: Array<keyof EarnParams> = ['from'];
@@ -58,20 +62,24 @@ const earningTypeLabelMap = {
   [YieldPoolType.SINGLE_FARMING]: 'single farming'
 };
 
-const Component = () => {
+const Component = ({ className }: ComponentProps) => {
   const { t } = useTranslation();
   const notify = useNotification();
   const { activeModal } = useContext(ModalContext);
   const { isWebUI } = useContext(ScreenContext);
   const { token } = useTheme() as Theme;
   const { setOnBack } = useContext(WebUIContext);
+  const navigate = useNavigate();
 
   const { closeAlert, defaultData, goBack, onDone,
     openAlert, persistData,
     setBackProps, setSubHeaderRightButtons } = useTransactionContext<EarnParams>();
 
-  const { redirectFromPreview, slug } = defaultData;
-
+  const { hasPreSelectTarget, redirectFromPreview, slug, target } = defaultData;
+  const defaultTarget = useRef<string>(target);
+  const autoCheckValidatorGetFromPreview = useRef<boolean>(true);
+  const autoCheckCompoundRef = useRef<boolean>(true);
+  const isReadyToShowAlertRef = useRef<boolean>(true);
   const { accounts, currentAccount, isAllAccount } = useSelector((state) => state.accountState);
   const chainInfoMap = useSelector((state) => state.chainStore.chainInfoMap);
   const poolInfoMap = useSelector((state) => state.earning.poolInfoMap);
@@ -102,6 +110,8 @@ const Component = () => {
   const { checkChainConnected, turnOnChain } = useChainConnection();
   const [isConnectingChainSuccess, setIsConnectingChainSuccess] = useState<boolean>(false);
   const [isLoadingChainConnection, setIsLoadingChainConnection] = useState<boolean>(false);
+  const [useParamValidator, setUseParamValidator] = useState<boolean>(hasPreSelectTarget);
+  const setSelectedAccountTypes = useSetSelectedAccountTypes(false);
 
   const poolInfo = poolInfoMap[slug] as YieldPoolInfo | undefined;
   const poolType = poolInfo?.type || '';
@@ -109,9 +119,11 @@ const Component = () => {
 
   const [isBalanceReady, setIsBalanceReady] = useState<boolean>(true);
   const [forceFetchValidator, setForceFetchValidator] = useState(false);
-  const [targetLoading, setTargetLoading] = useState(false);
+  const [targetLoading, setTargetLoading] = useState(true);
   const [stepLoading, setStepLoading] = useState<boolean>(true);
   const [screenLoading, setScreenLoading] = useState(true);
+  const [checkValidAccountLoading, setCheckValidAccountLoading] = useState(true);
+  const [checkCompoundLoading, setCheckCompoundLoading] = useState(true);
   const [submitString, setSubmitString] = useState<string | undefined>();
   const [connectionError, setConnectionError] = useState<string>();
   const [, setCanMint] = useState(false);
@@ -517,6 +529,38 @@ const Component = () => {
       }
     };
 
+    const maxCount = poolInfo?.statistic?.maxCandidatePerFarmer ?? 1;
+    const userSelectedPoolCount = poolTargetValue?.split(',').length ?? 1;
+    const label = getValidatorLabel(chainValue);
+
+    if (userSelectedPoolCount < maxCount && label === 'Validator') {
+      openAlert({
+        title: t('Pay attention!'),
+        content: t('You are recommended to choose {{maxCount}} validators to optimize your earnings. Do you wish to continue with {{userSelectedPoolCount}} validator{{x}}?', { replace: { maxCount, userSelectedPoolCount, x: userSelectedPoolCount === 1 ? '' : 's' } }),
+        okButton: {
+          text: t('Continue'),
+          onClick: () => {
+            closeAlert();
+            submitData(currentStep)
+              .catch(onError)
+              .finally(() => {
+                setSubmitLoading(false);
+              });
+          }
+        },
+        cancelButton: {
+          text: t('Go back'),
+          onClick: () => {
+            setSubmitLoading(false);
+            closeAlert();
+          }
+        },
+        closable: false
+      });
+
+      return;
+    }
+
     setTimeout(() => {
       submitData(currentStep)
         .catch(onError)
@@ -524,7 +568,7 @@ const Component = () => {
           setSubmitLoading(false);
         });
     }, 300);
-  }, [currentStep, onError, onSuccess, poolInfo, poolTargets, processState.feeStructure, processState.steps]);
+  }, [chainValue, closeAlert, currentStep, onError, onSuccess, openAlert, poolInfo, poolTargetValue, poolTargets, processState.feeStructure, processState.steps, t]);
 
   const renderMetaInfo = useCallback(() => {
     if (!poolInfo || !inputAsset) {
@@ -683,6 +727,201 @@ const Component = () => {
   }, [chainAsset, poolInfo]);
 
   useEffect(() => {
+    let timer: NodeJS.Timer;
+    let timeout: NodeJS.Timeout;
+
+    if (checkCompoundLoading && redirectFromPreview) {
+      const checkCompoundReady = () => {
+        if (compound) {
+          clearInterval(timer);
+          clearTimeout(timeout);
+          setCheckCompoundLoading(false);
+        }
+      };
+
+      timer = setInterval(checkCompoundReady, 500);
+
+      timeout = setTimeout(() => {
+        clearInterval(timer);
+        setCheckCompoundLoading(false);
+      }, 5000);
+    } else {
+      setTimeout(() => setCheckCompoundLoading(false), 350);
+    }
+
+    return () => {
+      clearInterval(timer);
+      clearTimeout(timeout);
+    };
+  }, [checkCompoundLoading, compound, redirectFromPreview]);
+
+  useEffect(() => {
+    if (redirectFromPreview && !accountSelectorList.length && checkValidAccountLoading) {
+      const isChainEvm = chainInfoMap[poolChain] && _isChainEvmCompatible(chainInfoMap[poolChain]);
+
+      setSelectedAccountTypes([isChainEvm ? EVM_ACCOUNT_TYPE : SUBSTRATE_ACCOUNT_TYPE]);
+      navigate('/home/earning', { state: { view: 'position', redirectFromPreview: true, chainName: chainInfoMap[poolChain]?.name || '' } });
+    } else {
+      setCheckValidAccountLoading(false);
+    }
+  }, [accountSelectorList, chainInfoMap, checkValidAccountLoading, navigate, poolChain, redirectFromPreview, setSelectedAccountTypes]);
+
+  const checkUnrecommendedValidator = useCallback((onValid?: () => void) => {
+    fetchPoolTarget({ slug }).then((rs) => {
+      const isValidatorSupported = rs.targets.some((item) => {
+        if (poolType === YieldPoolType.NOMINATION_POOL) {
+          return (item as NominationPoolInfo).id.toString() === defaultTarget.current;
+        } else if (poolType === YieldPoolType.NATIVE_STAKING) {
+          const _item = item as ValidatorInfo;
+          const key = getValidatorKey(_item.address, _item.identity);
+
+          return key === defaultTarget.current;
+        } else {
+          return false;
+        }
+      });
+
+      if (!isValidatorSupported && defaultTarget.current) {
+        isReadyToShowAlertRef.current && openAlert({
+          title: t('Unrecommended validator'),
+          type: NotificationType.ERROR,
+          content: t('Your chosen validator is not recommended by SubWallet as staking with this validator won’t accrue any rewards. Select another validator and try again.'),
+          cancelButton: {
+            text: t('Dismiss'),
+            onClick: () => {
+              isReadyToShowAlertRef.current = true;
+              closeAlert();
+            },
+            icon: XCircle
+          },
+          okButton: {
+            text: t('Select validators'),
+            onClick: () => {
+              isReadyToShowAlertRef.current = true;
+              closeAlert();
+              activeModal('target');
+            },
+            icon: CheckCircle
+          }
+        });
+        isReadyToShowAlertRef.current = false;
+      } else {
+        onValid && onValid();
+      }
+    }).catch((e) => console.error(e));
+  }, [activeModal, closeAlert, openAlert, poolType, slug, t]);
+
+  const isUnstakeAll = useMemo(() => {
+    if (compound) {
+      if (compound.nominations && compound.nominations.length) {
+        return compound.nominations.some((item) => item.activeStake === '0' && item.status === EarningStatus.NOT_EARNING);
+      } else {
+        return true;
+      }
+    }
+
+    return false;
+  }, [compound]);
+
+  useEffect(() => {
+    if (hasPreSelectTarget && !targetLoading && !screenLoading && !checkCompoundLoading) {
+      if (compound) {
+        if (autoCheckCompoundRef.current) {
+          autoCheckCompoundRef.current = false;
+
+          if (isUnstakeAll) {
+            if (poolType === YieldPoolType.NOMINATION_POOL) {
+              isReadyToShowAlertRef.current && openAlert({
+                title: t('Pay attention'),
+                content: t('This account is unstaking all stake and can\'t nominate validators. You can change your account on the Account tab or try again after withdrawing unstaked funds'),
+                type: NotificationType.WARNING,
+                okButton: {
+                  text: t('I understand'),
+                  onClick: () => {
+                    isReadyToShowAlertRef.current = true;
+                    closeAlert();
+                  },
+                  icon: CheckCircle
+                }
+              });
+              isReadyToShowAlertRef.current = false;
+
+              return;
+            } else if (poolType === YieldPoolType.NATIVE_STAKING) {
+              if (_STAKING_CHAIN_GROUP.para.includes(chainValue)) {
+                isReadyToShowAlertRef.current && openAlert({
+                  title: t('Pay attention'),
+                  content: t('This account is unstaking all stake and can\'t nominate validators. You can change your account on the Account tab or try again after withdrawing unstaked funds'),
+                  type: NotificationType.WARNING,
+                  okButton: {
+                    text: t('I understand'),
+                    onClick: () => {
+                      isReadyToShowAlertRef.current = true;
+                      closeAlert();
+                    },
+                    icon: CheckCircle
+                  }
+                });
+                isReadyToShowAlertRef.current = false;
+
+                return;
+              } else {
+                return;
+              }
+            }
+          }
+
+          const content = poolType === YieldPoolType.NATIVE_STAKING
+            ? t('This account is currently nominating {{number}} validators. You can change validators or change your account on the Account tab', { number: compound.nominations.length })
+            : poolType === YieldPoolType.NOMINATION_POOL
+              ? t('This account is currently a member of a nomination pool. You can continue using nomination pool, explore other Earning options or change your account on the Account tab', { x: compound?.nominations[0]?.validatorIdentity || compound?.nominations[0]?.validatorAddress || '' })
+              : '';
+
+          const onPressContinue = () => {
+            isReadyToShowAlertRef.current = true;
+
+            if (poolType === YieldPoolType.NATIVE_STAKING) {
+              checkUnrecommendedValidator(() => form.setFieldValue('target', defaultTarget.current));
+            }
+
+            closeAlert();
+          };
+
+          const onPressCancel = () => {
+            if (poolType === YieldPoolType.NOMINATION_POOL) {
+              goBack();
+            }
+
+            setUseParamValidator(false);
+            closeAlert();
+          };
+
+          isReadyToShowAlertRef.current && openAlert({
+            title: t('Pay attention'),
+            content: content,
+            className: CN(className, 'earning-alert-modal'),
+            type: NotificationType.WARNING,
+            cancelButton: {
+              text: poolType === YieldPoolType.NATIVE_STAKING ? t('Keep current validators') : poolType === YieldPoolType.NOMINATION_POOL ? t('Explore Earning options') : '',
+              onClick: onPressCancel
+            },
+            okButton: {
+              text: poolType === YieldPoolType.NATIVE_STAKING ? t('Change validators') : poolType === YieldPoolType.NOMINATION_POOL ? t('Use nomination pool') : '',
+              onClick: onPressContinue
+            }
+          });
+          isReadyToShowAlertRef.current = false;
+        }
+      } else {
+        if (autoCheckValidatorGetFromPreview.current) {
+          autoCheckValidatorGetFromPreview.current = false;
+          checkUnrecommendedValidator();
+        }
+      }
+    }
+  }, [isUnstakeAll, checkUnrecommendedValidator, className, closeAlert, compound, form, goBack, openAlert, poolType, hasPreSelectTarget, t, targetLoading, chainValue, screenLoading, checkCompoundLoading]);
+
+  useEffect(() => {
     if (poolChain) {
       if (altChain) {
         if (checkChainConnected(poolChain) && checkChainConnected(altChain)) {
@@ -765,12 +1004,12 @@ const Component = () => {
   }, [form, inputAsset?.slug]);
 
   useEffect(() => {
-    if (!fromValue) {
-      if ((redirectFromPreview && accountSelectorList.length >= 1) || accountSelectorList.length === 1) {
+    if (!fromValue && (isAllAccount || accountSelectorList.length === 1)) {
+      if ((hasPreSelectTarget && accountSelectorList.length >= 1) || accountSelectorList.length === 1) {
         form.setFieldValue('from', accountSelectorList[0].address);
       }
     }
-  }, [accountSelectorList, form, fromValue, redirectFromPreview]);
+  }, [accountSelectorList, form, fromValue, isAllAccount, hasPreSelectTarget]);
 
   useEffect(() => {
     if (currentStep === 0) {
@@ -930,22 +1169,34 @@ const Component = () => {
   const amountInputRef = form.getFieldInstance('value');
 
   useEffect(() => {
-    if (redirectFromPreview) {
+    if (hasPreSelectTarget) {
       // eslint-disable-next-line @typescript-eslint/no-unsafe-call,@typescript-eslint/no-unsafe-member-access
       amountInputRef?.focus?.();
     }
-  }, [amountInputRef, redirectFromPreview]);
+  }, [amountInputRef, hasPreSelectTarget]);
+
+  const validatorDefaultValue = (() => {
+    if (useParamValidator) {
+      return defaultData.target === 'not-support' ? '' : defaultTarget.current;
+    } else {
+      if (defaultData.target === 'not-support' || !!compound) {
+        return undefined;
+      } else {
+        return defaultTarget.current;
+      }
+    }
+  })();
 
   return (
     <>
       {
-        screenLoading && (
+        (screenLoading || checkValidAccountLoading) && (
           <LoadingScreen />
         )
       }
 
       {
-        !screenLoading && (
+        (!screenLoading && !checkValidAccountLoading) && (
           <>
             <div className={'__transaction-block'}>
               <TransactionContent>
@@ -1032,6 +1283,7 @@ const Component = () => {
                     >
                       <EarningPoolSelector
                         chain={poolChain}
+                        defaultValue={defaultData.target === 'not-support' || !!compound ? '' : defaultData.target}
                         disabled={submitLoading}
                         from={fromValue}
                         label={t('Select pool')}
@@ -1048,6 +1300,7 @@ const Component = () => {
                     >
                       <EarningValidatorSelector
                         chain={chainValue}
+                        defaultValue={validatorDefaultValue}
                         disabled={submitLoading}
                         from={fromValue}
                         loading={targetLoading}
@@ -1161,7 +1414,7 @@ const Wrapper: React.FC<Props> = (props: Props) => {
       path={'/transaction/earn'}
       stores={['price', 'chainStore', 'assetRegistry', 'earning']}
     >
-      <Component />
+      <Component className={className} />
     </EarnOutlet>
   );
 };
@@ -1175,6 +1428,28 @@ const Earn = styled(Wrapper)<Props>(({ theme: { token } }: Props) => {
     marginLeft: 'auto',
     marginRight: 'auto',
     gap: token.size,
+
+    '&.earning-alert-modal': {
+      alignItems: 'center',
+
+      '.ant-sw-modal-body': {
+        paddingBottom: token.padding
+      },
+
+      '.ant-sw-modal-footer': {
+        flexWrap: 'wrap',
+        gap: token.paddingSM
+      },
+
+      '.ant-btn.ant-btn.ant-btn': {
+        marginInlineStart: 0,
+        order: 2
+      },
+
+      '.ant-btn.ant-btn.ant-btn:last-child': {
+        order: 1
+      }
+    },
 
     '.__process-item-wrapper': {
       paddingBottom: token.paddingSM,