+++ packages/extension-web-ui/src/hooks/earning/useYieldGroupInfo.ts
@@ -1,15 +1,31 @@
 // Copyright 2019-2022 @polkadot/extension-web-ui authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
-import { _ChainAsset } from '@subwallet/chain-list/types';
+import { _ChainAsset, _MultiChainAsset } from '@subwallet/chain-list/types';
+import { _getAssetOriginChain } from '@subwallet/extension-base/services/chain-service/utils';
+import { _STAKING_CHAIN_GROUP } from '@subwallet/extension-base/services/earning-service/constants';
 import { calculateReward } from '@subwallet/extension-base/services/earning-service/utils';
-import { YieldPoolInfo } from '@subwallet/extension-base/types';
+import { YieldPoolInfo, YieldPoolType } from '@subwallet/extension-base/types';
 import { BN_TEN, BN_ZERO } from '@subwallet/extension-web-ui/constants';
 import { useAccountBalance, useGetChainSlugsByCurrentAccount, useSelector, useTokenGroup } from '@subwallet/extension-web-ui/hooks';
 import { BalanceValueInfo, YieldGroupInfo } from '@subwallet/extension-web-ui/types';
 import { isRelatedToAstar } from '@subwallet/extension-web-ui/utils';
 import BigN from 'bignumber.js';
 import { useMemo } from 'react';
+
+const isRelatedToRelayChain = (group: string, assetRegistry: Record<string, _ChainAsset>, multiChainAssetMap: Record<string, _MultiChainAsset>) => {
+  if (assetRegistry[group]) {
+    return _STAKING_CHAIN_GROUP.relay.includes(_getAssetOriginChain(assetRegistry[group]));
+  }
+
+  if (multiChainAssetMap[group]) {
+    const originChainAsset = multiChainAssetMap[group].originChainAsset;
+
+    return _STAKING_CHAIN_GROUP.relay.includes(_getAssetOriginChain(assetRegistry[originChainAsset]));
+  }
+
+  return false;
+};
 
 function calculateTotalValueStaked (poolInfo: YieldPoolInfo, assetRegistry: Record<string, _ChainAsset>, priceMap: Record<string, number>) {
   const asset = assetRegistry[poolInfo.metadata.inputAsset];
@@ -76,7 +92,14 @@ const useYieldGroupInfo = (): YieldGroupInfo[] => {
 
           exists.isTestnet = exists.isTestnet || chainInfo.isTestnet;
           exists.poolSlugs.push(pool.slug);
-          exists.totalValueStaked = exists.totalValueStaked.plus(calculateTotalValueStaked(pool, assetRegistry, priceMap));
+
+          if (exists.isRelatedToRelayChain) {
+            if (pool.type === YieldPoolType.NATIVE_STAKING) {
+              exists.totalValueStaked = calculateTotalValueStaked(pool, assetRegistry, priceMap);
+            }
+          } else {
+            exists.totalValueStaked = exists.totalValueStaked.plus(calculateTotalValueStaked(pool, assetRegistry, priceMap));
+          }
         } else {
           const token = multiChainAssetMap[group] || assetRegistry[group];
 
@@ -101,6 +124,8 @@ const useYieldGroupInfo = (): YieldGroupInfo[] => {
             apy = calculateReward(pool.statistic?.totalApr).apy;
           }
 
+          const checkRelatedRelaChain = isRelatedToRelayChain(group, assetRegistry, multiChainAssetMap);
+
           result[group] = {
             group: group,
             token: token.slug,
@@ -113,8 +138,11 @@ const useYieldGroupInfo = (): YieldGroupInfo[] => {
             poolListLength: 1,
             poolSlugs: [pool.slug],
             description: pool.metadata.description,
-            totalValueStaked: calculateTotalValueStaked(pool, assetRegistry, priceMap),
-            minJoin: pool.statistic?.earningThreshold?.join
+            totalValueStaked: (checkRelatedRelaChain && pool.type !== YieldPoolType.NATIVE_STAKING)
+              ? BN_ZERO
+              : calculateTotalValueStaked(pool, assetRegistry, priceMap),
+            minJoin: pool.statistic?.earningThreshold?.join,
+            isRelatedToRelayChain: checkRelatedRelaChain
           };
         }
       }